requirements.txt
pandas
streamlit
fastapi
uvicorn
scikit-learn
joblib
ruff
requests
pytest
httpx
sqlalchemy
psycopg2-binary

# Dependencies for Chatbot
google-generativeai
sentence-transformers
pgvector
python-dotenv
langchain
